render_requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
email-validator==2.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.25
yfinance==0.2.28
schedule==1.2.0
supabase==2.3.4
trafilatura==1.8.0